Edit bkgrnd, collider, & cam, add maze & materials
Changed background to solid color (black). Added possible maze design and frictionless materials to the walls and character to fix "getting stuck on walls" bug. Changed character box collider to a circle collider so the player more easily rolls over the edge of a wall. Added camera movement that follows the player.