Add timer and game over text
The player has 20 seconds to find an exit to the maze. If the timer runs out, the text "Game Over!" will be displayed, and the "You Win!" text game object is destroyed so that it isn't possible for the player to see that he/she won after the timer runs out. So, even if the exit is found after the time = 0, the player will not see that they won. Similarly, the "Game Over!" text object is destroyed if the exit is found before the timer runs out.