Add docs folder and status file
Leaderboard for lvl1

Current bug: When a new player entry is added to the table,
some overlapping of text occurs every once in a while.

Added new docs folder and status.txt file. status.txt is
stored in the docs folder and contains the console output
resulting from the 'git status' command.